@@ -9,11 +9,11 @@
     "@types/mongoose": "^5.11.96",
     "@types/node": "^20.11.5",
     "body-parser": "^2.2.0",
+    "bootstrap-icons": "^1.13.1",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
     "mongodb": "^4.5.0",
-    "mongoose": "^8.14.1",
-    "typescript": "^5.2.2"
+    "mongoose": "^8.14.1"
   },
   "scripts": {
     "test": "mocha -r ts-node/register test/**/*.test.ts",
@@ -36,12 +36,13 @@
     "chai": "^5.2.0",
     "copyfiles": "^2.4.1",
     "eslint": "^9.26.0",
-    "mocha": "^11.3.0",
+    "mocha": "^11.4.0",
     "proxyquire": "^2.1.3",
     "rimraf": "^6.0.1",
     "sinon": "^20.0.0",
     "supertest": "^7.1.1",
     "ts-node": "^10.9.2",
+    "typescript": "^5.8.3",
     "typescript-eslint": "^8.32.1"
   }
 }
