@@ -1,26 +1,30 @@
+// This script is used to create sample data for the MongoDB database.
 
 db = db.getSiblingDB('tutorialPlatform')
 
 db.createCollection('tutorials');
 db.createCollection('comments');
+db.createCollection('communityNotes');
 
 tutorials = db.getCollection('tutorials');
 comments = db.getCollection('comments');
+communityNotes = db.getCollection('communityNotes')
 
 // Clear data
 tutorials.remove({})
 comments.remove({});
+communityNotes.remove({})
 
 // Insert Tutorials 
 tutorialList = [
   {
     title: 'Getting Started with MongoDB',
-    description: 'Learn the basics of MongoDB database operations and queries.',
+    text: 'Learn the basics of MongoDB database operations and queries.',
     tutorialId: 'tut001',
     createdDate: new Date('2025-01-10'),
     updatedDate: new Date('2025-01-15'),
     authorId: 'auth001',
-    authorName: 'John Smith',
+    authorName: 'diy_life98',
     category: 'Database',
     tags: ['MongoDB', 'NoSQL', 'Database'],
     views: 1240,
@@ -29,33 +33,35 @@ tutorialList = [
     steps: [
       {
         stepNumber: 1,
-        description: 'Install MongoDB Community Edition on your machine',
-        imageUrls: ['mongodb_install.jpg'],
+        title: 'Install MongoDB Community Edition on your machine',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       },
       {
         stepNumber: 2,
-        description: 'Start the MongoDB server using the mongod command',
-        imageUrls: ['mongodb_start.jpg'],
+        title: 'Start the MongoDB server using the mongod command',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       },
       {
         stepNumber: 3,
-        description: 'Connect to MongoDB using the MongoDB shell',
-        imageUrls: ['mongodb_connect.jpg'],
-        videoUrls: ['mongodb_connection_video.mp4']
+        title: 'Connect to MongoDB using the MongoDB shell',
+        imageUrls: [],
+        videoUrls: []
       }
     ],
     published: true
   },
   {
     title: 'Building a RESTful API with Node.js and Express',
-    description: 'Create a complete RESTful API using Node.js and Express framework.',
+    text: 'Create a complete RESTful API using Node.js and Express framework.',
     tutorialId: 'tut002',
     createdDate: new Date('2025-02-05'),
     updatedDate: new Date('2025-02-12'),
     authorId: 'auth002',
-    authorName: 'Sarah Johnson',
+    authorName: 'quickFix01',
     category: 'Web Development',
     tags: ['Node.js', 'Express', 'API', 'JavaScript'],
     views: 985,
@@ -64,26 +70,30 @@ tutorialList = [
     steps: [
       {
         stepNumber: 1,
-        description: 'Set up your Node.js project and install Express',
-        imageUrls: ['node_setup.jpg'],
+        title: 'Set up your Node.js project and install Express',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       },
       {
         stepNumber: 2,
-        description: 'Create your first Express server',
-        imageUrls: ['express_server.jpg'],
+        title: 'Create your first Express server',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       },
       {
         stepNumber: 3,
-        description: 'Implement CRUD operations for your API',
+        title: 'Implement CRUD operations for your API',
+        text: 'Placeholder text',
         imageUrls: ['crud_operations.jpg'],
-        videoUrls: ['express_crud_video.mp4']
+        videoUrls: []
       },
       {
         stepNumber: 4,
-        description: 'Add authentication to your API',
-        imageUrls: ['api_auth.jpg'],
+        title: 'Add authentication to your API',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       }
     ],
@@ -91,12 +101,12 @@ tutorialList = [
   },
   {
     title: 'Introduction to React Hooks',
-    description: 'Learn how to use React Hooks to build dynamic user interfaces.',
+    text: 'Learn how to use React Hooks to build dynamic user interfaces.',
     tutorialId: 'tut003',
     createdDate: new Date('2025-03-01'),
     updatedDate: new Date('2025-03-10'),
     authorId: 'auth001',
-    authorName: 'John Smith',
+    authorName: 'diy_life98',
     category: 'Frontend',
     tags: ['React', 'JavaScript', 'Hooks', 'Frontend'],
     views: 1567,
@@ -105,26 +115,30 @@ tutorialList = [
     steps: [
       {
         stepNumber: 1,
-        description: 'Understanding React Hooks and their benefits',
+        title: 'Understanding React Hooks and their benefits',
+        text: 'Placeholder text',
         imageUrls: ['hooks_intro.jpg'],
         videoUrls: []
       },
       {
         stepNumber: 2,
-        description: 'Using the useState Hook',
-        imageUrls: ['usestate_hook.jpg'],
-        videoUrls: ['usestate_demo.mp4']
+        title: 'Using the useState Hook',
+        text: 'Placeholder text',
+        imageUrls: [],
+        videoUrls: []
       },
       {
         stepNumber: 3,
-        description: 'Using the useEffect Hook',
-        imageUrls: ['useeffect_hook.jpg'],
+        title: 'Using the useEffect Hook',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       },
       {
         stepNumber: 4,
-        description: 'Creating custom Hooks',
-        imageUrls: ['custom_hooks.jpg'],
+        title: 'Creating custom Hooks',
+        text: 'Placeholder text',
+        imageUrls: [],
         videoUrls: []
       }
     ],
@@ -138,7 +152,7 @@ tutResult = tutorials.insertMany(tutorialList);
 commentList = [
   {
     commentId: 'comm001',
-    noteId: 'tut001',
+    noteId: 'note001',
     userId: 'user001',
     text: 'Great tutorial! The step-by-step explanation made it easy to follow.',
     votesUp: 5,
@@ -147,7 +161,7 @@ commentList = [
   },
   {
     commentId: 'comm002',
-    noteId: 'tut001',
+    noteId: 'note001',
     userId: 'user002',
     text: "I'm having trouble with step 3. Could you provide more details on the connection string format?",
     votesUp: 2,
@@ -156,7 +170,7 @@ commentList = [
   },
   {
     commentId: 'comm003',
-    noteId: 'tut001',
+    noteId: 'note002',
     userId: 'auth001',
     text: 'Sure! The connection string format is: mongodb://localhost:27017/yourDatabaseName',
     votesUp: 8,
@@ -165,7 +179,7 @@ commentList = [
   },
   {
     commentId: 'comm004',
-    noteId: 'tut001',
+    noteId: 'note002',
     userId: 'user003',
     text: "I think there's a mistake in step 2. The command should be 'mongod --dbpath /path/to/data' to specify the data directory.",
     votesUp: 15,
@@ -174,7 +188,7 @@ commentList = [
   },
   {
     commentId: 'comm005',
-    noteId: 'tut002',
+    noteId: 'note003',
     userId: 'user004',
     text: 'Excellent tutorial on Express! I found the authentication section particularly helpful.',
     votesUp: 7,
@@ -183,7 +197,7 @@ commentList = [
   },
   {
     commentId: 'comm006',
-    noteId: 'tut002',
+    noteId: 'note002',
     userId: 'user005',
     text: 'This tutorial really helped me understand RESTful APIs. Thanks!',
     votesUp: 4,
@@ -192,7 +206,7 @@ commentList = [
   },
   {
     commentId: 'comm007',
-    noteId: 'tut003',
+    noteId: 'note002',
     userId: 'user006',
     text: 'The explanation of useEffect was very clear. I finally understand how to handle side effects properly!',
     votesUp: 9,
@@ -201,7 +215,7 @@ commentList = [
   },
   {
     commentId: 'comm008',
-    noteId: 'tut003',
+    noteId: 'note001',
     userId: 'user007',
     text: "I'd recommend adding a section on the useContext hook as well. It's quite important for state management.",
     votesUp: 12,
@@ -212,6 +226,44 @@ commentList = [
 
 comResult = comments.insertMany(commentList);
 
+// Insert Community Notes
+communityNoteList = [
+  {
+    noteId: 'note001',
+    tutorialId: 'tut001',
+    userId: 'user008',
+    title: 'For MacOS users',
+    text: 'Heads up: on macOS you may need to brew-services start mongodb before step 2.',
+    votesUp: 14,
+    votesDown: 1,
+    createdDate: new Date('2025-01-18'),
+    updatedDate: new Date('2025-01-19')
+  },
+  {
+    noteId: 'note002',
+    tutorialId: 'tut001',
+    userId: 'user009',
+    title: 'Suggesting JWT',
+    text: 'In step 4, consider using JWT expiry of 1h instead of default 24h for better security.',
+    votesUp: 9,
+    votesDown: 0,
+    createdDate:new Date('2025-02-13')
+  },
+  {
+    noteId: 'note003',
+    tutorialId: 'tut002',
+    userId: 'user010',
+    title: 'Request for a example/sandbox',
+    text: 'It would be nice to show a code sandbox link for live editing of your custom Hooks example.',
+    votesUp: 22,
+    votesDown: 2,
+    createdDate: new Date('2025-03-08'),
+    updatedDate: new Date('2025-03-09')
+  }
+]
+
+cnResult = communityNotes.insertMany(communityNoteList);
+
 tutorials.createIndex({ tutorialId: 1 }, { unique: true });
 tutorials.createIndex({ authorId: 1 });
 tutorials.createIndex({ category: 1 });
@@ -220,5 +272,7 @@ tutorials.createIndex({ published: 1 });
 
 comments.createIndex({ tutorialId: 1 });
 comments.createIndex({ userId: 1 });
-comments.createIndex({ parentCommentId: 1 });
-comments.createIndex({ isAmendment: 1 });
+
+communityNotes.createIndex({ noteId: 1 }, { unique: true })
+communityNotes.createIndex({ tutorialId: 1 })
+communityNotes.createIndex({ userId: 1 })
